rename: http 와 관련된 클래스를 http 패키지로 이동
